fix: cli options namespace clash with custom rocksdb fix/rocksdb-plugin
In order to gain access to the Blockchain data we override the default storage service with a
customised version of the rocksdb plugin that allows us to re-use the same underlying key/value
storage objects and get around the issues with creating a second instance of rocksdb backed by
the same files.

However the newer version of picocli throws an exception when an attempt is made to register the
same cli options. The fix was to create separately namespaced options.

The implications are that when using exflo user's will need to use our namespace rocksdb options
instead of the standard besu rocksdb cli options.