feat: simplified key/value storage interception
I realised we didn't need a copy of the rocksdb plugin, we simply needed to be a
bit more intelligent how about how we interacted with the storage service. The
new approach will allow multiple exflo plugins to run at the same time and will
not require custom exflo options for controlling rocksdb.